laurel_sprout: props: Remove forced disablement of compr_voip audio f… …eature
Three of our platforms - Yoshino, Nile and Ganges - use this feature.
This is not only evident from the fact that CAF audio enables it on
sdm660 and msm8998, but the lack of `echo-reference-voip` mixer paths.
Only echo-reference is available, where `-voip` is appened if compr_voip
is disabled.

Explicitly enabling this feature (which defaults to false) in platform
repos fixes issues with calls that utilize echo-reference.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>
Signed-off-by: Hemant Sachdeva <hemant.evolver@gmail.com>